requirements.txt
langchain==0.2.1
langchain_community==0.2.1
langchain_core==0.2.3
python-dotenv==1.0.1
